Finally fixed the unittest for full coverage and simplified the checking for eligible multi-part files.

+++ the_ark/s3_client.py
@@ -63,29 +63,27 @@ class S3Client(object):
             mime_type = mimetypes.guess_type(filename) if mime_type is None else mime_type
             s3_file.set_metadata('Content-Type', mime_type)
 
-            # - Check if file is a buffer or disk file
-            if type(file_to_store) == str:
-                # - Check if file that is getting uploaded is greater than chunk_at_size then upload cool multi style
-                if os.path.getsize(file_to_store) > chunk_at_size:
-                    file_count = 0
-                    # - Split the file and get it chunky
-                    split_file_dir = self._split_file(file_to_store)
+            # - Check if file is a buffer or disk file and if file that is getting uploaded is greater than
+            # chunk_at_size then upload cool multi style
+            if type(file_to_store) == str and os.path.getsize(file_to_store) > chunk_at_size:
+                file_count = 0
+                # - Split the file and get it chunky
+                split_file_dir = self._split_file(file_to_store)
 
-                    # - Initiate the file to be uploaded in parts
-                    multipart_file = self.bucket.initiate_multipart_upload(key_name=s3_file.key,
-                                                                           metadata=s3_file.metadata)
+                # - Initiate the file to be uploaded in parts
+                multipart_file = self.bucket.initiate_multipart_upload(key_name=s3_file.key, metadata=s3_file.metadata)
 
-                    # - Upload the file parts
-                    for files in os.listdir(split_file_dir):
-                        file_count += 1
-                        file_part = open(os.path.join(split_file_dir, files), 'rb')
-                        multipart_file.upload_part_from_file(file_part, file_count)
+                # - Upload the file parts
+                for files in os.listdir(split_file_dir):
+                    file_count += 1
+                    file_part = open(os.path.join(split_file_dir, files), 'rb')
+                    multipart_file.upload_part_from_file(file_part, file_count)
 
-                    # - Complete the upload
-                    multipart_file.complete_upload()
+                # - Complete the upload
+                multipart_file.complete_upload()
 
-                    # - Remove the folder from splitting the file
-                    shutil.rmtree(split_file_dir)
+                # - Remove the folder from splitting the file
+                shutil.rmtree(split_file_dir)
 
             # - Upload the file as a whole
             else:
